+++ src/main/java/com/inksetter/twist/expression/IntegerLiteral.java
@@ -3,17 +3,19 @@ package com.inksetter.twist.expression;
 import com.inksetter.twist.EvalContext;
 import com.inksetter.twist.Expression;
 
+import java.math.BigInteger;
+
 public class IntegerLiteral implements Expression {
-    public IntegerLiteral(Integer value) {
+    public IntegerLiteral(BigInteger value) {
         _value = value;
     }
 
-    public Integer evaluate(EvalContext ctx) {
+    public BigInteger evaluate(EvalContext ctx) {
         return _value;
     }
 
     @Override
     public String toString() { return _value.toString(); }
 
-    private final Integer _value;
+    private final BigInteger _value;
 }
